+++ src/time_series/mod.rs
@@ -5,5 +5,6 @@
 
 pub mod cyclical;
 pub mod diff;
+pub mod expanding;
 pub mod lag;
 pub mod rolling;